% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTijd_style.R
\name{theme_DeTijd}
\alias{theme_DeTijd}
\title{Add De Tijd theme to ggplot chart}
\usage{
theme_DeTijd()
}
\description{
This function allows you to add the De Tijd theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_DeTijd()
}
\keyword{theme_DeTijd}
